Upstream version 0.7

+++ src/rt/rust_uv.cpp
@@ -229,7 +229,7 @@ rust_uv_timer_init(uv_loop_t* loop, uv_timer_t* timer) {
 
 extern "C" int
 rust_uv_timer_start(uv_timer_t* the_timer, uv_timer_cb cb,
-                        uint32_t timeout, uint32_t repeat) {
+                    int64_t timeout, int64_t repeat) {
     return uv_timer_start(the_timer, cb, timeout, repeat);
 }
 
@@ -401,8 +401,7 @@ rust_uv_get_data_for_uv_handle(uv_handle_t* handle) {
 }
 
 extern "C" void
-rust_uv_set_data_for_uv_handle(uv_handle_t* handle,
-        void* data) {
+rust_uv_set_data_for_uv_handle(uv_handle_t* handle, void* data) {
     handle->data = data;
 }
